% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputreat.R
\name{imputreat}
\alias{imputreat}
\title{Auxiliary function for the Difference-in-Differences with bad covariates.}
\usage{
imputreat(dt, imput_var, group_var, id_var, time_var)
}
\arguments{
\item{dt}{A dataset object, in "long" format.}

\item{imput_var}{String name of the imputation variable.}

\item{group_var}{String name of the group variable.}

\item{id_var}{String name of the individual identifier variable.}

\item{time_var}{String name of the time identifier variable.}
}
\value{
A modified data.table object with new column for the imputation variable.
}
\description{
Auxiliary function for the Difference-in-Differences with bad covariates.
}
\examples{
# Assuming `data` is a data.table with appropriate columns:
imputed_data <- imputreat(data, "imput_var_name", "group_var_name", 
                          "id_var_name", "time_var_name")

}
\references{
Caetano, C., & Callaway, B., & Payne, S., & Sant'Anna, H.(2024). Difference-in-differences With Bad Covariates.
}
\author{
Stroud Payne (1) and Hugo Sant'Anna (2)
(1) Vanderbilt University
(2) University of Georgia
}
